requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
blue==0.9.0
click==8.1.3
decorator==5.1.1
Django==4.0.6
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
executing==0.8.3
Faker==13.15.0
flake8==4.0.1
ipdb==0.13.9
ipython==8.4.0
isort==5.10.1
jedi==0.18.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.12.0
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
toml==0.10.2
tomli==2.0.1
traitlets==5.3.0
wcwidth==0.2.5
